refactor(roadmaps): relocate constants, update imports and raise node detail limit
