+++ apps/api/src/modules/roadmaps/constants/roadmap.constants.ts
@@ -14,7 +14,7 @@ export const QUIZ_GENERATION_POLL_TIMEOUT_MS = 45_000;
 export const QUIZ_GENERATION_POLL_INTERVAL_MS = 1_500;
 export const QUIZ_PASSING_SCORE_PCT = 60;
 export const QUIZ_REVIEW_SUGGESTION = 'You should review this part before continuing.';
-export const NODE_DETAIL_RESOURCE_LIMIT = 2;
+export const NODE_DETAIL_RESOURCE_LIMIT = 5;
 export const MILESTONE_PASS_THRESHOLD_PCT = 80;
 export const MILESTONE_TEST_SUITE_CASE_COUNT = 6;
 export const MILESTONE_TEST_SUITE_POLL_TIMEOUT_MS = 45_000;
