+++ apps/api/src/modules/roadmaps/services/roadmap-milestone.service.ts
@@ -28,15 +28,6 @@ import type {
   MilestoneTestSuiteRecord,
 } from '../utils/roadmap-records';
 
-import { MilestoneExecutionClient } from '../milestone-execution.client';
-import { delay } from '../utils/async';
-import {
-  assertMilestoneSubmissionPayload,
-  sanitizeMilestoneOutputLog,
-  toMilestoneTestResult,
-} from '../utils/milestone-output';
-import { getRoadmapRelationAccessWhere } from '../utils/roadmap-access';
-import { formatMilestoneSubmission, formatMilestoneTestSuite } from '../utils/roadmap-formatters';
 import {
   MILESTONE_EXECUTION_TIMEOUT_MS,
   MILESTONE_GENERATED_TEST_COMMAND,
@@ -47,7 +38,16 @@ import {
   MILESTONE_TEST_SUITE_POLL_INTERVAL_MS,
   MILESTONE_TEST_SUITE_POLL_TIMEOUT_MS,
   MILESTONE_TEST_SUITE_SELECT,
-} from '../utils/roadmap.constants';
+} from '../constants/roadmap.constants';
+import { MilestoneExecutionClient } from '../milestone-execution.client';
+import { delay } from '../utils/async';
+import {
+  assertMilestoneSubmissionPayload,
+  sanitizeMilestoneOutputLog,
+  toMilestoneTestResult,
+} from '../utils/milestone-output';
+import { getRoadmapRelationAccessWhere } from '../utils/roadmap-access';
+import { formatMilestoneSubmission, formatMilestoneTestSuite } from '../utils/roadmap-formatters';
 import { RoadmapProgressService } from './roadmap-progress.service';
 
 @Injectable()
